chore: refactored client
